+++ changelog.md
@@ -1,3 +1,9 @@
+## April 24th 2019 - 0.8.0
+
+- markRead now supports sending a message_id to mark the channel read up to (and including) that specific message
+- added markAllRead client method
+- countUnread can be called without any parameters now client-side and it will default to current user's read state
+
 ## April 9th 2019 - 0.7.2
 
 - queryChannels used to return the list of members twice, this has now been resolved. However if you were using the duplicate list of members in channel.members you'll want to update to Object.values(channel.state.members)